requirements.txt
discord>=2.3.2
Aiohttp
psutil
dotenv
bs4
python-dateutil
playwright
fastapi
uvicorn
requests
